backend/requirements.txt
cosdata-sdk
sentence-transformers
fastapi
uvicorn
python-dotenv
requests
numpy